requirements.txt
opencv-python
numpy
pillow
ascii_magic
mediapipe
yagmail
pyzbar
qrcode